+++ src/Attributes/Response206Attribute.cs
@@ -6,6 +6,7 @@ namespace PowerUtils.AspNetCore.Attributes
 {
     public class Response206Attribute : ProducesResponseTypeAttribute
     {
+        public Response206Attribute() : base(StatusCodes.Status206PartialContent) { }
         public Response206Attribute(Type type) : base(type, StatusCodes.Status206PartialContent) { }
     }
 }