+++ tests/PowerUtils.AspNetCore.Extensions.Tests/AttributesTests/ResponseTypes2XXTests.cs
@@ -6,7 +6,22 @@ namespace PowerUtils.AspNetCore.Extensions.Tests.AttributesTests;
 public class ResponseTypes2XXTests
 {
     [Fact]
-    public void Constructor_Response200Attribute()
+    public void Constructor_WithoutType_Response200Attribute()
+    {
+        // Arrange && Act
+        var act = new Response200Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(200);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response200Attribute()
     {
         // Arrange && Act
         var act = new Response200Attribute(typeof(FakeResponse));
@@ -21,7 +36,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response201Attribute()
+    public void Constructor_WithoutType_Response201Attribute()
+    {
+        // Arrange && Act
+        var act = new Response201Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(201);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response201Attribute()
     {
         // Arrange && Act
         var act = new Response201Attribute(typeof(FakeResponse));
@@ -36,7 +66,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response202Attribute()
+    public void Constructor_WithoutType_Response202Attribute()
+    {
+        // Arrange && Act
+        var act = new Response202Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(202);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response202Attribute()
     {
         // Arrange && Act
         var act = new Response202Attribute(typeof(FakeResponse));
@@ -51,7 +96,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response203Attribute()
+    public void Constructor_WithoutType_Response203Attribute()
+    {
+        // Arrange && Act
+        var act = new Response203Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(203);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response203Attribute()
     {
         // Arrange && Act
         var act = new Response203Attribute(typeof(FakeResponse));
@@ -66,7 +126,7 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response204Attribute()
+    public void Constructor_WithoutType_Response204Attribute()
     {
         // Arrange && Act
         var act = new Response204Attribute();
@@ -81,7 +141,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response205Attribute()
+    public void Constructor_WithoutType_Response205Attribute()
+    {
+        // Arrange && Act
+        var act = new Response205Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(205);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response205Attribute()
     {
         // Arrange && Act
         var act = new Response205Attribute(typeof(FakeResponse));
@@ -96,7 +171,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response206Attribute()
+    public void Constructor_WithoutType_Response206Attribute()
+    {
+        // Arrange && Act
+        var act = new Response206Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(206);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response206Attribute()
     {
         // Arrange && Act
         var act = new Response206Attribute(typeof(FakeResponse));
@@ -111,7 +201,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response207Attribute()
+    public void Constructor_WithoutType_Response207Attribute()
+    {
+        // Arrange && Act
+        var act = new Response207Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(207);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response207Attribute()
     {
         // Arrange && Act
         var act = new Response207Attribute(typeof(FakeResponse));
@@ -126,7 +231,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response208Attribute()
+    public void Constructor_WithoutType_Response208Attribute()
+    {
+        // Arrange && Act
+        var act = new Response208Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(208);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response208Attribute()
     {
         // Arrange && Act
         var act = new Response208Attribute(typeof(FakeResponse));
@@ -141,7 +261,22 @@ public class ResponseTypes2XXTests
     }
 
     [Fact]
-    public void Constructor_Response226Attribute()
+    public void Constructor_WithoutType_Response226Attribute()
+    {
+        // Arrange && Act
+        var act = new Response226Attribute();
+
+
+        // Assert
+        act.StatusCode.Should()
+            .Be(226);
+
+        act.Type.Should()
+            .Be(typeof(void));
+    }
+
+    [Fact]
+    public void Constructor_WithType_Response226Attribute()
     {
         // Arrange && Act
         var act = new Response226Attribute(typeof(FakeResponse));